refactor(tui): retain incremental canvas budget
Fold incremental terminal painting into the existing canvas module after Work Control added a new production helper on dev.
